Servlet: flush before start writing, lazy flush in copy loop
Flush before start writing to make status code and headers available
earlier so users don't need to flush response channel.

Make flush in copy loop only if there are no outstanding bytes
in the response byte channel otherwise go to the next copy
iteration as we know that is is not going to suspend
so servlet writer will do flush() later when the byte channel
get drained.