Validate scheme and contents in authorization header
This change adds validation for the scheme and contents when trying to
parse the HTTP Basic credentials from the Authorization HTTP header.
This prevents the HTTP Basic protected routes from crashing when the
user sends an Authorization header with a scheme other than "Basic" or
when the user sends a malformed Base64 string.