Fixed issue #827: WebSocket masking cannot be enabled
The bug is caused by the use of overriding to set the masking property to true, which has no effect on the actual session delegate. 
An init {} block should be used to initialise the "masking" property to true. This also preserves the delegation mechanism to work properly. Changing the masking value of the wrapper class will also be propagated to the underlying delegate.