Fix Apache client engine freeze at request initiation
When sending a request, it is likely that consumeContent callback will
be triggered before inDecoderThread invocation, especially on Windows.
Usually it is not a problem since we do requestInput. However, if the response
is short or empty or it is the last chunk, then there will be no more callback invocations so there will be no
change for rendezvous. So to avoid this unfortunate case we introduce decoder
reference that is populated in the callback in case when no suspension yet
and then we pick it up in decoding loop initiation later.