Fix websocket flush implementation
It is a common pattern to send a close frame and then do flush.
However, the web socket writer loop coroutine does close the queue channel
when receiving a close frame so a flush request may be discarded
that is unfortunate. So, instead of a crash, we simply wait for a write loop completion in flush function.

The fact that we do not rethrow a ClosedSendChannelException is not an issue
since we are waiting for the loop completion.
Also, such behaviour is more convenient, so users will not get flacky exceptions.