Move the DefaultByteBufferPool to ktor-network module (#1629)
* Move the DefaultByteBufferPool to ktor-network module

The DefaultByteBufferPool lived in the ktor-http-cio module, but was used by the ktor-network-tls module. This meant the the ktor-network-tls module had both the ktor-http-cio and ktor-http modules as dependencies. This is undesirable as the HTTP code is not required for TLS or plain-socket connections, and it increases the size of applications, as well as adding unnecessary items to the classpath.

The ktor-network already had a file, Pools.kt, for declaring an ObjectPool for UDP datagrams, so the general purpose pool from the ktor-http-cio module has been moved to that file. A further benefit is that there is now only 1 implementation of DirectByteBufferPool in the ktor project.

After moving the ObjectPool to the ktor-network module, the ktor-network module no longer uses anything from ktor-http-cio and the dependency was replaced with the ktor-utils module.

Remove KDoc suppression since KDoc is provided