set AuthScheme.OAuth to `Bearer` as defined in RFC6750
In RFC6750 the auth-scheme is defined as `Bearer` (not `OAuth`). See https://tools.ietf.org/html/rfc6750#section-3:

> All challenges defined by this specification MUST use the auth-scheme value "Bearer".

This is also commonly used by servers which respond with `Bearer` as well so this must match to be used by an (not yet implemented) `OAuthAuthProvider` (see https://github.com/ktorio/ktor/issues/1595).