Add locale to ThymeleafContent
#### Subsystem
ktor-thymeleaf
#### Motivation
In order to support i18n in templates, Thymeleaf provides accessors to resource bundles based on a set locale. When this feature's process method is called a new Context object is created with an empty constructor ```(Thymeleaf.kt:84)```, which then initializes a locale property to the default locale of the current jvm instance. Ktor provides ```call.request.acceptedLanguageItems()``` to access the ```accepted-languages``` field of the http header, and it would be helpful to use a locale based on these requested languages for template rendering.
#### Solution
As the ThymeleafContent object is the public way to provide content from Ktor to Thymeleaf using this feature, an optional locale parameter is added to the ThymeleafContent constructor. If a locale is provided, the Thymeleaf context object is constructed with it. If a locale is not provided, the old behavior is preserved by using a default value of ```Locale.getDefault()```. Previously omitting a locale in the context constructor would initialize a locale with ```Locale.getDefault()``` in the context's super class ```(org.thymeleaf.context.AbstractContext:62)```.
### Example
#### Routing
```
get("/demo") {
 val locale = Locale("es") // default locale is English
 call.respond(
 ThymeleafContent(
 template = "demo",
 model = mapOf(),
 locale = locale
 )
 )
}
```
#### demo.html
```
<!DOCTYPE html>
<html xmlns:th="http://www.thymeleaf.org">
<head><meta charset="UTF-8"></head>
 <body>
 <h2 th:text="#{greeting}"></h2>
 </body>
</html>
```
#### demo_en.properties
```
greeting=Hello, World!
```
#### demo_es.properties
```
greeting=¡Hola, Mundo!
```
#### Result
```
<!DOCTYPE html>
<html>
<head><meta charset="UTF-8"></head>
 <body>
 <h2>¡Hola, Mundo!</h2>
 </body>
</html>
```