Test i18n template rendering
Adds a test for template rendering when providing a locale.

By providing a Locale range to the Accept-Language header in the request, the server is able to construct an appropriate Locale to pass to the Thymeleaf context. The use of this Locale is then used to select the appropriate properties file to fill in the template.