Add Moshi support
This is a proposal PR with Moshi support. Moshi is a spiritual successor to Gson, faster, actively maintained, and has true first first-party support for Kotlin (both via reflect artifact or code gen). Two of its original authors are also two of the three maintainers of Gson as well.

Its use is mostly similar to gson, but one significant caveat is that unlike gson it will not try to auto-magically fall back to Map/Set/List adapters if it receives concrete implementations. Normally moshi handles this by requesting the explicit adapter type, but Ktor does not currently offer a way to indicate the encoding type for serialization (only offers the object, with encoders expected to introspect it via reflection).