Jackson: Do not convert to String before converting to JSON
The existing implementation of the JacksonConverter used a TextContent,
which meant that the ObjectMapper had to serialize the entire output
value to a String before responding.

For large output objects, this causes significant allocation preassure.
It can avoid by instead streaming the output. This change does that.

The change is fairly small, simply using different existing methods on
the ObjectMapper in combination with a OutputStreamContent.

There is an added test to verify that it got UTF-16 conversion right.