KTOR-5023 Make certificate generation helpers more flexible (#3208)
* Add tests for certificate generation

* Add missing documentation on keysGenerationAlgorithm function

* Make certificate() reusable and de-duplicate CertificateBuilder code

* Make domains and IP addresses configurable in CertificateBuilder

* Avoid deprecated Date types and use Duration API to make it public-friendly

* Only pass the PublicKey instead of KeyPair when the private key is not needed

The creation of the certificate itself is only about the public key that is
being certified. We only need the corresponding private key in the context
of key stores, and thus we should express that in the types.

Note that the signer key is different, because we need the private key to
sign the certificate, and the public key to verify the signature, so it makes
sense to pass the whole key pair.

* Switch to built-in X500Principal type for a more public-friendly API

* Delete unused helper functions

* Rename and document generateX509Certificate

* Make keyType configurable in CertificateBuilder

* Provide default algorithms for more convenient use of CertificateBuilder

These defaults are already used in the publicly advertised generateCertificate
functions, which could be completely replaced by this builder.

* Generalize CertificateBuilder to more than self-signed certificates

This makes the builder general enough to be reused from both
generateCertificate functions (self-signed or CA-signed), and even
makes them almost obsolete.