Fix NullPointerException when clearToken is called on AuthTokenHolder while suspended
`loadToken` and `setToken` are suspending functions that call a suspending lambda at some point during their execution.

If `clearToken` is called while they are waiting for that suspending lamba call to return, the code crashes with a `NullPointerException` when we force unwrap `refreshTokensDeferred.value` or `loadTokensDeferred.value`.

One solution to avoid the crash would be to simply not force unwrap and use the `?` notation. However, that also means that if there are several callers waiting for `loadToken()` to complete, they will be suspended forever. Only the 1st caller of `loadToken()` will get its value back.

So I opted for capturing the deferred value when the function starts and calling `.complete(newTokens)` on it instead. This ensures all function calls will eventually complete.