Fix out-of-bounds reads in testBigFrame
testBigFrame always wrote the whole backing array of the ByteBuffer to
the websocket. However, this buffer has always been overallocated - the
serializer never fully filled it.

This leads to an issue where using the backing array of the buffer,
includes a bunch of null bytes at the end, which were written to the
websocket unintentionally. This should've always broken the websocket,
however due to a lack of validation of incoming socket frames this issue
was never exposed.

This is a test-only issue.