KTOR-6221: Fix reduced concurrent reqs in Apache5 (#3738)
When using the Apache5 engine, total concurrent requests to a single
route were limited to 5 requests. This is due to the code which tried to
increase the concurrency to the ktor-standard 1000 concurrent having a
typo and setting the total max connections twice and missing the max
connections
completely.

(cherry picked from commit fa5cba1012302e9ab431826cebd8a4eef7929e1e)