Allow `respondTemplate` set HTTP status (#3162)
* Allow `respondTemplate` set HTTP status

The default HTTP status in `io.ktor.server.engine.BaseApplicationResponse#commitHeaders` is `HttpStatusCode.OK` (200). This is almost always applicable when responding with a template. However, in the case of templates being used along with the `StatusPages` plugin or for response entities for HTTP methods, it is necessary to respond with an appropriate HTTP status code.

This change allows for an optional parameter for the caller to set the status code along with responding with a template. IIUC, it should be binary compatible.

---------

Co-authored-by: leonid.stashevsky <leonid.stashevsky@gmail.com>