Use KotlinHierarchyTemplate for structuring source sets
KotlinHierarchyTemplate allows writing KMP source set structures with less code
and without explicit dependsOn relations.

There is one downside is that all source sets are created now if at least
one of the target is declared. i.e. for example if JS target is declared
but wasm is not. Then jsAndWasmSharedMain source set will be created anyway.

It also fixes the issue with Kotlin 2.0 KT-68212
In :ktor-network jvmAndNixTest has an expect that in K1 was actualized in iosTest
there was no direct or transitive dependsOn relation from iosTest to jvmAndNixTest
and thus in K2 such actualization was prohibited. In this commit Source Set
structure is sanitized and now dependsOn relations are in good order.

^KT-68212 Fixed