Lift test dependency to the source set that uses it
CIOHttpServerTest extends from the base class
HttpServerCommonTestSuite. It in turn extends EngineTestBase
declared in :ktor-server:ktor-server-test-base. Dependency
on this module is declared for jvmTest instead of the shared
jvmAndNixTest that contains CIOHttpServerTest. The missing
dependency on a transitive module with the grandparent class
results in a MISSING_DEPENDENCY_SUPERCLASS in K2 IDE.

KTIJ-28250