Resolve a classpath clash WA
The class `org.jetbrains.kotlin.konan.util.DependencyDirectories` is present in both `org.jetbrains.kotlin:kotlin-compiler-embeddable` and `org.jetbrains.kotlin:kotlin-native-utils`. Previously KGP had a dependency to both of them. Now we intentionally removed the dependency on `kotlin-compiler-embeddable` and it’s working fine in most of setups. But here the problem is that kotlinter plugin adds a dependency on `org.jetbrains.kotlin:kotlin-compiler-embeddable:1.8.20` that was bumped to the actual Kotlin version, but not anymore. The possibility of such particular classpath clashes will be completely removed in KT-70247 when we stop depending on compiler modules from KGP. It will be partially resolved in KT-70251.