KTOR-8409 Fix closing of OkHttp SSE connection
Currently, when the outer coroutine context is canceled, the OkHttp
connection thread is leaked. This is because the connection is never
canceled when the session job is completed. This is also because the
OkHttp SSE session doesn't have it's own Job for use within the SSE
session block and was defaulting to the open call request