KTOR-6198: Fix incorrect CloseReason in Client/WebSocket/Darwin
Don't cancel the WebSocket job if receiveMessage() fails with
exception and the WebSocket is closed. When the WebSocket is closed
the callback method URLSession:webSocketTask:didCloseWithCode:reason:
in the Ktor library takes care of canceling the WebSocket job after
the close code is propagated to the client application.
The exception in receiveMessage() was canceling the WebSocket job
and because of that the close code and reason were not propagated.